Move elastic search config closer to its tests
