+++ sagan-site/src/it/java/sagan/search/support/InMemoryElasticSearchConfig.java
@@ -1,4 +1,4 @@
-package saganx;
+package sagan.search.support;
 
 import sagan.search.support.SearchService;
